Product Single Order Information

+++ Modules/Shop/Traits/CartTraits.php
@@ -104,7 +104,7 @@ trait CartTraits
 
     private function clear_session_cart($session)
     {
-        return CartRecord::where('session_id', $session)->delete();
+        return CartRecord::where('session_id', $session)->orWhere('user_id', auth()->id())->delete();
     }
 
     private function clear_all_cart($session)
